feat: add navigation
