+++ example/package.json
@@ -17,10 +17,13 @@
     "@pagopa/io-react-native-login-utils": "^1.0.1",
     "@pagopa/react-native-cie": "^1.3.0",
     "@react-native-async-storage/async-storage": "^1.24.0",
+    "@react-navigation/native": "^6.1.18",
+    "@react-navigation/native-stack": "^6.11.0",
     "@reduxjs/toolkit": "^2.2.7",
     "react": "18.2.0",
     "react-native": "0.72.14",
-    "react-native-safe-area-context": "^4.9.0",
+    "react-native-safe-area-context": "^4.10.8",
+    "react-native-screens": "^3.34.0",
     "react-native-webview": "^13.10.5",
     "react-redux": "^9.1.2",
     "redux": "^5.0.1",
