+++ example/src/App.tsx
@@ -1,20 +1,21 @@
 import React from "react";
 import "react-native-url-polyfill/auto";
 import { SafeAreaProvider } from "react-native-safe-area-context";
-
+import { NavigationContainer } from "@react-navigation/native";
 import { Provider } from "react-redux";
-
-import MainComponent from "./MainComponent";
 import { PersistGate } from "redux-persist/integration/react";
 import { store, persistor } from "./store/store";
+import { MainStackNavigator } from "./navigator/MainStackNavigator";
 
 export default function App() {
   return (
     <Provider store={store}>
       <PersistGate loading={null} persistor={persistor}>
-        <SafeAreaProvider>
-          <MainComponent />
-        </SafeAreaProvider>
+        <NavigationContainer>
+          <SafeAreaProvider>
+            <MainStackNavigator />
+          </SafeAreaProvider>
+        </NavigationContainer>
       </PersistGate>
     </Provider>
   );
