+++ example/src/scenarios/get-pid.ts
@@ -15,8 +15,8 @@ import {
 import uuid from "react-native-uuid";
 import { generate } from "@pagopa/io-react-native-crypto";
 import { Alert } from "react-native";
-import type { PidContext } from "../MainComponent";
 import appFetch from "../utils/fetch";
+import type { PidContext } from "../screens/HomeScreen";
 
 /**
  * Callback used to set the PID and its crypto context in the app state which is later used to obtain a credential
