+++ package.json
@@ -28,7 +28,7 @@
   ],
   "scripts": {
     "test": "jest",
-    "typecheck": "tsc --noEmit",
+    "tsc": "tsc --noEmit",
     "lint": "eslint \"src/**\" \"example/src/**\" \"__mocks__/**\"",
     "prepack": "bob build",
     "release": "release-it",
